generate query calls as lens logic
