+++ src/lib/codegen/components/common.rs
@@ -243,9 +243,6 @@ pub trait ComponentManifest: std::fmt::Debug {
     fn user_impl_required(&self) -> bool;
     fn generate_user_impl_template(&self) -> anyhow::Result<TokenStream>;
     fn get_sources(&self) -> Sources;
-    fn default_query_identifier(&self) -> Option<&str> {
-        Option::None
-    }
     fn dependencies(&self) -> Vec<String> {
         vec![]
     }
