+++ src/lib/codegen/components/snapshot_indexer.rs
@@ -134,9 +134,6 @@ impl ComponentManifest for SnapshotIndexerComponentManifest {
             attributes: attr,
         }
     }
-    fn default_query_identifier(&self) -> Option<&str> {
-        Option::Some("get_last_snapshot_value : () -> (text)")
-    }
     fn custom_tags(&self) -> HashMap<String, String> {
         let mut res = HashMap::new();
         let (interval_key, interval_val) = custom_tags_interval_sec(self.interval);
